Added AudioPlayer class
Added a thread whose job it is to play audio files or text to speech. By having it be a single thread, I don't have to worry about playing more than one thing at a time, and I can also interrupt it anytime I want. There may be a problem with overwhelming the input_queue, though.