Big Changes to Tattle Core
I should have made these more atomic, but it boils down to this: I don't care about a dial tone! The whole reason I made this branch: forget it. Instead we got to the core of what I was doing in the first place: recording tattling. And this update does it.

We can now record tattles right the disk! And the menu system works! Feels pretty good.