Reworked TattleCore into a class
It just feels cleaner to me to have member variables instead of big
globals. So made it a class, made playback its own function. I'll
probably do the same for voice recording, I like the idea of the
functions performing and just returning the next state we should be in.
Although, even that is too much, probably.